restructured components into folders for styling
