+++ src/App.css
@@ -7,6 +7,12 @@
   pointer-events: none;
 }
 
+/* .card {
+  width: 100%;
+  margin-left: 10px;
+  margin-right: auto;
+} */
+
 @media (prefers-reduced-motion: no-preference) {
   .App-logo {
     animation: App-logo-spin infinite 20s linear;
