+++ src/components/Table/index.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import "./style.css";
 
 function Table(props) {
   const { employees } = props;
